Sort out remaining types
