+++ packages/types/src/discord/guilds.ts
@@ -1,6 +1,6 @@
 /** Types for: https://discord.com/developers/docs/resources/guild */
 
-import type { DiscordChannel } from './channels.js'
+import type { ChannelTypes, DiscordChannel, DiscordOverwrite, SortOrderTypes } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
 import type { DiscordPresenceUpdate } from './gateway.js'
 import type { OAuth2Scope } from './oauth2.js'
@@ -598,14 +598,52 @@ export enum DiscordGuildOnboardingPromptType {
   DropDown,
 }
 
-/** https://discord.com/developers/docs/resources/guild#modify-guild-welcome-screen */
-export interface DiscordModifyGuildWelcomeScreen {
-  /** Whether the welcome screen is enabled */
-  enabled?: boolean | null
-  /** Channels linked in the welcome screen and their display options */
-  welcome_screen?: DiscordWelcomeScreenChannel[] | null
-  /** The server description to show in the welcome screen */
-  description?: string | null
+/** https://discord.com/developers/docs/resources/guild#create-guild-channel-json-params */
+export interface DiscordCreateGuildChannel {
+  /** Channel name (1-100 characters) */
+  name: string
+  /** The type of channel */
+  type?: ChannelTypes
+  /** Channel topic (0-1024 characters) */
+  topic?: string
+  /** The bitrate (in bits) of the voice channel (voice only) */
+  bitrate?: number
+  /** The user limit of the voice channel (voice only) */
+  user_limit?: number
+  /** Amount of seconds a user has to wait before sending another message (0-21600); bots, as well as users with the permission `manage_messages` or `manage_channel`, are unaffected */
+  rate_limit_per_user?: number
+  /** Sorting position of the channel */
+  position?: number
+  /** The channel's permission overwrites */
+  permission_overwrites?: DiscordOverwrite[]
+  /** Id of the parent category for a channel */
+  parent_id?: string
+  /** Whether the channel is nsfw */
+  nsfw?: boolean
+  /** Default duration (in minutes) that clients (not the API) use for newly created threads in this channel, to determine when to automatically archive the thread after the last activity */
+  default_auto_archive_duration?: number
+  /** Emoji to show in the add reaction button on a thread in a forum channel */
+  default_reaction_emoji?: {
+    /** The id of a guild's custom emoji. Exactly one of `emojiId` and `emojiName` must be set. */
+    emoji_id?: string | null
+    /** The unicode character of the emoji. Exactly one of `emojiId` and `emojiName` must be set. */
+    emoji_name?: string | null
+  }
+  /** Set of tags that can be used in a forum channel */
+  available_tags?: Array<{
+    /** The id of the tag */
+    id: string
+    /** The name of the tag (0-20 characters) */
+    name: string
+    /** whether this tag can only be added to or removed from threads by a member with the MANAGE_THREADS permission */
+    moderated: boolean
+    /** The id of a guild's custom emoji */
+    emoji_id: string
+    /** The unicode character of the emoji */
+    emoji_name?: string
+  }>
+  /** the default sort order type used to order posts in forum channels */
+  default_sort_order?: SortOrderTypes | null
 }
 
 /** https://discord.com/developers/docs/resources/guild#modify-guild-channel-positions */
@@ -626,4 +664,26 @@ export interface DiscordBulkBan {
   banned_users: string[]
   /** list of user ids, that were not banned */
   failed_users: string[]
+}
+
+/** https://discord.com/developers/docs/resources/guild#get-guild-prune-count */
+export interface DiscordPrunedCount {
+  pruned: number
+}
+
+/** https://discord.com/developers/docs/resources/guild#get-guild-vanity-url */
+// TODO: This should provably have another name for clarity to what it really rappresent
+export interface DiscordVanityUrl {
+  code: string | null
+  uses: number
+}
+
+/** https://discord.com/developers/docs/resources/guild#modify-guild-welcome-screen */
+export interface DiscordModifyGuildWelcomeScreen {
+  /** Whether the welcome screen is enabled */
+  enabled?: boolean | null
+  /** Channels linked in the welcome screen and their display options */
+  welcome_screen?: DiscordWelcomeScreenChannel[] | null
+  /** The server description to show in the welcome screen */
+  description?: string | null
 }
