+++ packages/types/src/index.ts
@@ -1,4 +1,3 @@
-export * from './discord.js'
 export * from './discord/applications.js'
 export * from './discord/auditLogs.js'
 export * from './discord/autoModeration.js'
