+++ packages/types/src/shared.ts
@@ -1,86 +1,4 @@
-import type { Locales } from './discord/reference.js'
-
 export type BigString = bigint | string
-
-export type GatewayDispatchEventNames =
-  | 'READY'
-  | 'APPLICATION_COMMAND_PERMISSIONS_UPDATE'
-  | 'AUTO_MODERATION_RULE_CREATE'
-  | 'AUTO_MODERATION_RULE_UPDATE'
-  | 'AUTO_MODERATION_RULE_DELETE'
-  | 'AUTO_MODERATION_ACTION_EXECUTION'
-  | 'CHANNEL_CREATE'
-  | 'CHANNEL_UPDATE'
-  | 'CHANNEL_DELETE'
-  | 'CHANNEL_PINS_UPDATE'
-  | 'THREAD_CREATE'
-  | 'THREAD_UPDATE'
-  | 'THREAD_DELETE'
-  | 'THREAD_LIST_SYNC'
-  | 'THREAD_MEMBER_UPDATE'
-  | 'THREAD_MEMBERS_UPDATE'
-  | 'GUILD_AUDIT_LOG_ENTRY_CREATE'
-  | 'GUILD_CREATE'
-  | 'GUILD_UPDATE'
-  | 'GUILD_DELETE'
-  | 'GUILD_BAN_ADD'
-  | 'GUILD_BAN_REMOVE'
-  | 'GUILD_EMOJIS_UPDATE'
-  | 'GUILD_STICKERS_UPDATE'
-  | 'GUILD_INTEGRATIONS_UPDATE'
-  | 'GUILD_MEMBER_ADD'
-  | 'GUILD_MEMBER_REMOVE'
-  | 'GUILD_MEMBER_UPDATE'
-  | 'GUILD_MEMBERS_CHUNK'
-  | 'GUILD_ROLE_CREATE'
-  | 'GUILD_ROLE_UPDATE'
-  | 'GUILD_ROLE_DELETE'
-  | 'GUILD_SCHEDULED_EVENT_CREATE'
-  | 'GUILD_SCHEDULED_EVENT_UPDATE'
-  | 'GUILD_SCHEDULED_EVENT_DELETE'
-  | 'GUILD_SCHEDULED_EVENT_USER_ADD'
-  | 'GUILD_SCHEDULED_EVENT_USER_REMOVE'
-  | 'GUILD_SOUNDBOARD_SOUND_CREATE'
-  | 'GUILD_SOUNDBOARD_SOUND_UPDATE'
-  | 'GUILD_SOUNDBOARD_SOUND_DELETE'
-  | 'GUILD_SOUNDBOARD_SOUNDS_UPDATE'
-  | 'SOUNDBOARD_SOUNDS'
-  | 'INTEGRATION_CREATE'
-  | 'INTEGRATION_UPDATE'
-  | 'INTEGRATION_DELETE'
-  | 'INTERACTION_CREATE'
-  | 'INVITE_CREATE'
-  | 'INVITE_DELETE'
-  | 'MESSAGE_CREATE'
-  | 'MESSAGE_UPDATE'
-  | 'MESSAGE_DELETE'
-  | 'MESSAGE_DELETE_BULK'
-  | 'MESSAGE_REACTION_ADD'
-  | 'MESSAGE_REACTION_REMOVE'
-  | 'MESSAGE_REACTION_REMOVE_ALL'
-  | 'MESSAGE_REACTION_REMOVE_EMOJI'
-  | 'PRESENCE_UPDATE'
-  | 'STAGE_INSTANCE_CREATE'
-  | 'STAGE_INSTANCE_UPDATE'
-  | 'STAGE_INSTANCE_DELETE'
-  | 'TYPING_START'
-  | 'USER_UPDATE'
-  | 'VOICE_CHANNEL_EFFECT_SEND'
-  | 'VOICE_STATE_UPDATE'
-  | 'VOICE_SERVER_UPDATE'
-  | 'WEBHOOKS_UPDATE'
-  | 'ENTITLEMENT_CREATE'
-  | 'ENTITLEMENT_UPDATE'
-  | 'ENTITLEMENT_DELETE'
-  | 'SUBSCRIPTION_CREATE'
-  | 'SUBSCRIPTION_UPDATE'
-  | 'SUBSCRIPTION_DELETE'
-  | 'MESSAGE_POLL_VOTE_ADD'
-  | 'MESSAGE_POLL_VOTE_REMOVE'
-
-export type GatewayEventNames = GatewayDispatchEventNames | 'RESUMED'
-
-export type Localization = Partial<Record<Locales, string>>
 
 export type AtLeastOne<T, U = { [K in keyof T]: Pick<T, K> }> = Partial<T> & U[keyof U]
 export type CamelCase<S extends string> = S extends `${infer T}_${infer U}` ? `${T}${Capitalize<CamelCase<U>>}` : S
